Move dispatcher and confirmer into batcher
